fix: return sha512t24u to default instead of CRC64, add md5 to prots
For the refseq data `hasher(f"{assembly_id}___{locus_id}")` produced 10s of thousands of collisions when using CRC64. I never saw that with sha512t24u.

The reason to switch to CRC64 was to allow API compatibility with UniProt (which uses CRC64 and md5) It turns out there are a number of collisions in uniprot with crc64 but not with md5, sha512, sha512t24u, or seguid. See the following for details:

https://gist.github.com/chasemc/74292ae0808f2a9fb7843ebec4b93bda

So, now sha512t24u is default again and md5 is always added, to maintain the ability to get a uniprot sequence by hash